Add check validation step to maas role
Validation is meant to verify that for each check file applied there
is a check generated. It must rely on each host having the correct
check files applied as the alternative to this would mean duplicating
significant logic.

For filesystem checks, the file on the filesystem will use a '.'
instead of a '/' but otherwise each check file in the filesystem
shall have a name that matches the label of the check as recorded
server-side.

The rpc-maas-tool has added a checks mode and a --tabs boolean option
to enable this validation

Conflicts:
	rpcd/playbooks/roles/rpc_maas/defaults/main.yml

(cherry picked from commit 5aef967c7004fefcb62589b1faf913484e9c2467)
(cherry picked from commit 5b089eb5d4e55a45bdbc2ca06312fa643b6d0700)