Ensure raxmon-agent's conf directory exists
When we're setting up the rackspace-monitoring-agent we should ensure
the conf.d directory exists. This should be created as part of the
package install but we found this was an issue on a torn down lab and
prevented a complete upgrade test. This could be an issue on other
environments too during upgrades.

Addresses #544